+++ firebase-vscode/src/test/utils/page_objects/execution.ts
@@ -14,11 +14,25 @@ export class ExecutionPanel {
     );
   }
 
+  async getVariables(): Promise<string> {
+    return this.runInConfigurationContext(async (configs) => {
+      return configs.variablesTextarea.getValue();
+    });
+  }
+
   async setVariables(variables: string): Promise<void> {
     // TODO revert to the original value after test
 
     await this.runInConfigurationContext(async (configs) => {
       await configs.variablesTextarea.setValue(variables);
+    });
+  }
+
+  async clickRerun(): Promise<void> {
+    return this.runInConfigurationContext(async (configs) => {
+      const rerunButton = await configs.rerunButton;
+      await rerunButton.waitForClickable();
+      await rerunButton.doubleClick(); // double click first transitions focus to window instead of notifs
     });
   }
 
@@ -42,6 +56,10 @@ export class ConfigurationView {
 
   get variablesTextarea() {
     return this.variablesView.$("textarea");
+  }
+
+  get rerunButton() {
+    return this.variablesView.$("vscode-button");
   }
 }