+++ firebase-vscode/CHANGELOG.md
@@ -1,5 +1,12 @@
 ## NEXT
 
+- [Added] Added rerun execution button in variables context
+- [Added] Provide default required variables during execution
+
+## 0.14.2
+
+- Updated internal `firebase-tools` dependency to 13.34.0
+
 ## 0.14.1
 
 - Updated internal `firebase-tools` dependency to 13.33.0